Change algorithm functions state methods
