+++ trace-to-events/src/channels/algorithm_states/differential.rs
@@ -1,10 +1,10 @@
 //! Provides objects for persisting state for the differential detector algorithm.
+use digital_muon_common::{Intensity, Time};
+
 use crate::{
-    parameters::{DifferentialThresholdDiscriminatorParameters, PeakHeightBasis, PeakHeightMode},
-    pulse_detection::{
-        detectors::differential_threshold_detector::DifferentialThresholdParameters,
-        window::FiniteDifferences,
-    },
+    channels::algorithm_states::AlgorithmState, parameters::{DifferentialThresholdDiscriminatorParameters, PeakHeightBasis, PeakHeightMode}, pulse_detection::{
+        EventsIterable as _, Real, WindowIterable, detectors::differential_threshold_detector::{DifferentialThresholdDetector, DifferentialThresholdParameters}, window::FiniteDifferences
+    }
 };
 
 /// Encapsulates settings to determine how peak heights should be calculated.
@@ -43,5 +43,44 @@ impl DifferentialThresholdDiscriminatorState {
                 basis: parameters.peak_height_basis.clone(),
             },
         }
+    }
+}
+
+impl AlgorithmState for DifferentialThresholdDiscriminatorState {
+    #[tracing::instrument(skip_all, level = "trace")]
+    fn find_events(
+        &mut self,
+        trace: impl Clone + ExactSizeIterator<Item = Real> + DoubleEndedIterator,
+        sample_time: Real,
+        polarity_sign: Real,
+        baseline: Real,
+    ) -> (Vec<Time>, Vec<Intensity>) {
+        let raw = trace.enumerate().map(|(i, v)| {
+            (
+                i as Real * sample_time,
+                polarity_sign * (v as Real - baseline),
+            )
+        });
+
+        let pulses = raw
+            .clone()
+            .window(self.finite_differences.clone_only_coefficients())
+            .events(DifferentialThresholdDetector::new(
+                &self.parameters,
+                self.peak_height.mode.clone(),
+            ));
+
+        let mut time = Vec::<Time>::new();
+        let mut voltage = Vec::<Intensity>::new();
+        for pulse in pulses {
+            time.push(pulse.0 as Time);
+            voltage.push(match self.peak_height.basis {
+                PeakHeightBasis::TraceBaseline => pulse.1.peak_height as Intensity,
+                PeakHeightBasis::PulseBaseline => {
+                    (pulse.1.peak_height - pulse.1.base_height) as Intensity
+                }
+            });
+        }
+        (time, voltage)
     }
 }
