+++ trace-to-events/src/channels/algorithm_states/smoothing.rs
@@ -1,10 +1,10 @@
 //! Provides objects for persisting state for the second-order smoothing algorithm.
+use digital_muon_common::{Intensity, Time};
+
 use crate::{
-    parameters::SmoothingDetectorParameters,
-    pulse_detection::{
-        Real,
-        window::convolution_filter::{ConvolutionFilter, KernelType},
-    },
+    channels::algorithm_states::AlgorithmState, parameters::SmoothingDetectorParameters, pulse_detection::{
+        EventsIterable, Real, detectors::{local_arg_min_detector::LocalArgMinDetector, region_detector::RegionDetector}, iterators::PaddingIterable, utils::{global_arg_min, std_dev}, window::{SliceWindow, convolution_filter::{ConvolutionFilter, KernelType}}
+    }
 };
 
 /// Encapsulates all settings and objects in the smoothing algorithm which persist across digitiser messages.
@@ -30,6 +30,73 @@ impl SmoothingDetectorState {
             }),
             cache: Default::default(),
         }
+    }
+}
+
+impl AlgorithmState for SmoothingDetectorState {
+    #[tracing::instrument(skip_all, level = "trace", fields(std_dev, num_regions))]
+    fn find_events(
+        &mut self,
+        trace: impl Clone + ExactSizeIterator<Item = Real> + DoubleEndedIterator,
+        sample_time: Real,
+        polarity_sign: Real,
+        baseline: Real,
+    ) -> (Vec<Time>, Vec<Intensity>) {
+        self.cache.ensure_time_data_written(trace.len(), sample_time);
+        // Get the radius of the kernel by right-bitshifting the size by one
+        // i.e. divide by 2, and round-down.
+        let kernel_radius = self.fin_diff_gaussian.kernel_size() >> 1;
+        let padded = trace
+            .clone()
+            .map(|v| polarity_sign * (v as Real - baseline))
+            .pad_reflect(kernel_radius, kernel_radius);
+        self.cache.ensure_cache_lengths(trace.len() + self.fin_diff_gaussian.kernel_size(), trace.len());
+        self.cache.write_input_values(padded);
+
+        self.fin_diff_gaussian.apply_to_slice(
+            self.cache.input_values.as_slice(),
+            self.cache.output_values.as_mut_slice(),
+        );
+
+        let percentile = ((trace.len() as Real * self.parameters.noise_centile) / 100.0) as usize;
+        let noise_std = std_dev(&self.cache.output_values[percentile..])
+            .expect("std_dev should exist, this should never fail.");
+        tracing::Span::current().record("std_dev", noise_std);
+
+        let output_iter = self.cache.output_values.iter().cloned().enumerate();
+
+        let regions = output_iter
+            .clone()
+            .events(RegionDetector::new(
+                -noise_std * self.parameters.nsig_noise,
+                self.parameters.min_size,
+            ))
+            .collect::<Vec<_>>();
+        tracing::Span::current().record("num_regions", regions.len());
+
+        let pulses = regions
+            .into_iter()
+            .flat_map(|region| {
+                let region_iter = output_iter.clone().take(region.1).skip(region.0);
+                if let Some(use_local_for_sizes_ge) = self.parameters.use_local_for_sizes_ge
+                    && region_iter.len() >= use_local_for_sizes_ge
+                {
+                    region_iter
+                        .events(LocalArgMinDetector::default())
+                        .collect::<Vec<_>>()
+                } else {
+                    vec![global_arg_min(region_iter)]
+                }
+            })
+            .collect::<Vec<_>>();
+
+        let mut times = Vec::<Time>::new();
+        let mut voltages = Vec::<Intensity>::new();
+        for time in pulses {
+            times.push(time as Time);
+            voltages.push(trace.clone().nth(time).expect("") as Intensity);
+        }
+        (times, voltages)
     }
 }
