+++ trace-to-events/src/channels/mod.rs
@@ -1,6 +1,5 @@
 //! Provides functions and structs which extract and return lists of muon events using specified detectors and settings.
 mod algorithm_states;
-mod algorithms;
 mod channel_state;
 
 pub(crate) use algorithm_states::LayerProcessingSettings;
